+++ WorldCities/WorldCitiesAPI/Data/GraphQL/Query.cs
@@ -1,4 +1,5 @@
-﻿using WorldCitiesAPI.Data.Models;
+﻿using Microsoft.EntityFrameworkCore;
+using WorldCitiesAPI.Data.Models;
 
 namespace WorldCitiesAPI.Data.GraphQL
 {
@@ -25,5 +26,47 @@ namespace WorldCitiesAPI.Data.GraphQL
     [UseFiltering]
     [UseSorting]
     public IQueryable<Country> GetCountries([Service] ApplicationDbContext context) => context.Countries;
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="context"></param>
+    /// <param name="pageIndex"></param>
+    /// <param name="pageSize"></param>
+    /// <param name="sortColumn"></param>
+    /// <param name="sortOrder"></param>
+    /// <param name="filterColumn"></param>
+    /// <param name="filterQuery"></param>
+    /// <returns></returns>
+    [Serial]
+    public async Task<ApiResult<CityDTO>> GetCitiesApiResult(
+      [Service] ApplicationDbContext context,
+      int pageIndex = 0,
+      int pageSize = 10,
+      string? sortColumn = null, 
+      string? sortOrder = null, 
+      string? filterColumn = null, 
+      string? filterQuery = null
+    )
+    {
+      return await ApiResult<CityDTO>.CreateAsync(
+        context.Cities.AsNoTracking()
+        .Select(c => new CityDTO() {
+          Id = c.Id,
+          Name = c.Name,
+          Lat = c.Lat,
+          Lon = c.Lon,
+          CountryId = c.Country!.Id,
+          CountryName = c.Country!.Name
+        }),
+        pageIndex,
+        pageSize,
+        sortColumn,
+        sortOrder,
+        filterColumn,
+        filterQuery
+      );
+    }
+
   }
 }